requirements.txt
kucoin-python==1.0.14
psycopg2==2.9.3
uuid==1.30
